Project_agent_app/requirements.txt
streamlit
langchain
langgraph
langchain-google-genai
sentence-transformers
requests
python-dotenv
endee
numpy